moved config logic to seperate file
